requirements.txt
stomp.py
pydantic
pydantic-settings
celery
redis
requests
